requirements.txt
customtkinter
python-dotenv
openai
requests
supabase
pillow
pydub
aiohttp
tenacity
discord.py
